move UniformArray class
